Small fixes after API V2 update
The updated oura_query.py script did not print out the results if it was
run independently. I brought back the output.

The fetch_data function in oura_query.py overrode the datatype to always
be "daily_sleep". This caused some issues when trying to push values to
the db, as the "daily_sleep" did not for example contain the "bedtime_end"
data. I updated it not to override the given datatype.

Oura has updated how they report the temperature deviations. They report
that in the daily_readiness now. Because of this, I decided to change the
default value for datatype to be "daily_readiness". This means that the
example usecase for fetching your temperatures remains exactly the same.

Updated the output of --help for oura_query.py to follow the changes.